+++ CHANGES.rst
@@ -9,6 +9,9 @@ https://github.com/zopefoundation/Zope/blob/4.x/CHANGES.rst
 5.0a1 (unreleased)
 ------------------
 
+- Fix ``OFS.Image.File.__str__`` for ``Pdata`` contents
+  (`#711 <https://github.com/zopefoundation/Zope/issues/711>`_)
+
 - Improve documentation for Zope's error logging services.
 
 Backwards incompatible changes